+++ src/pages/myLinks/ui/LinkCard/LinkCard.stories.tsx
@@ -7,9 +7,8 @@ const meta: Meta<LinkCardProps> = {
   component: LinkCard,
   tags: ['autodocs'],
   args: {
-    label: '합정동 모각디',
-    link: 'here.is.link.as.well',
-    onCopy: () => console.log('copied'),
+    name: '합정동 모각디',
+    url: 'here.is.link.as.well',
   },
 };
 
@@ -23,15 +22,10 @@ export const Showcase: Story = {
     <div
       style={{ display: 'flex', flexDirection: 'column', gap: 16, width: 350 }}
     >
+      <LinkCard name="합정동 모각디" url="here.is.link.as.well" />
       <LinkCard
-        label="합정동 모각디"
-        link="here.is.link.as.well"
-        onCopy={() => {}}
-      />
-      <LinkCard
-        label="이름도 길고 링크도 정말정말정말정말정말정말 길어요"
-        link="https://www.very-long-settlement-link.com/path/to/resource?id=12345678"
-        onCopy={() => {}}
+        name="이름도 길고 링크도 정말정말정말정말정말정말 길어요"
+        url="https://www.very-long-settlement-link.com/path/to/resource?id=12345678"
       />
     </div>
   ),
